fix: the messages are displayed on catch errors
