--- src/handlers/proxy.ts
@@ -1,30 +1,0 @@
-import { Context, Env, SupportedEventsU } from "../types";
-import { userSelfAssign, userStartStop } from "./user-start-stop";
-
-export enum HttpStatusCode {
-  OK = 200,
-  NOT_MODIFIED = 304,
-}
-
-export interface Result {
-  status: HttpStatusCode;
-  content?: string;
-  reason?: string;
-}
-
-const callbacks: { [k in SupportedEventsU]: (context: Context, env: Env) => Result | Promise<Result> } = {
-  "issues.assigned": userSelfAssign,
-  "issue_comment.created": userStartStop,
-};
-
-export function proxyCallbacks({ logger }: Context) {
-  return new Proxy(callbacks, {
-    get(target, prop: SupportedEventsU) {
-      if (!(prop in target)) {
-        logger.error(`${prop} is not supported, skipping.`);
-        return async () => ({ status: "skipped", reason: "unsupported_event" });
-      }
-      return target[prop].bind(target);
-    },
-  });
-}
