+++ src/handlers/result-types.ts
@@ -1,0 +1,10 @@
+export enum HttpStatusCode {
+  OK = 200,
+  NOT_MODIFIED = 304,
+}
+
+export interface Result {
+  status: HttpStatusCode;
+  content?: string;
+  reason?: string;
+}
